ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.44.0
openai==1.12.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.6.0
python-dotenv==1.0.0
numpy==1.26.3
